RemoteAuth に切り出した
IndexedDB で認証状態を保持するので、RemoteFirestore の挙動に問題はなかった
