+++ .env
@@ -1,2 +1,3 @@
+VITE_AUTH_EMULATOR=
 VITE_DB_VERSION=dev
 VITE_FIRESTORE_EMULATOR=
